+++ qudas/qudata/__init__.py
@@ -1,9 +1,7 @@
 # qudata/__init__.py
 
-# QuData, QuDataInput, QuDataOutput, QuDataBase クラスを外部から直接インポートできるようにする
+# QuData, QuDataBase クラスを外部から直接インポートできるようにする
 from .qudata import QuData
-from .qudata_input import QuDataInput
-from .qudata_output import QuDataOutput
 from .qudata_base import QuDataBase
 
-__all__ = ['QuData', 'QuDataInput', 'QuDataOutput', 'QuDataBase']
+__all__ = ['QuData', 'QuDataBase']