+++ test/test_qudata.py
@@ -13,7 +13,11 @@ import dimod
 from sympy import Symbol, symbols, lambdify
 from scipy.optimize import minimize, Bounds
 from datetime import timedelta
+from dotenv import load_dotenv
 import os
+
+# 環境変数の読み込み
+load_dotenv()
 
 
 def dicts_are_equal(dict1, dict2):
@@ -133,7 +137,7 @@ class TestQudata(unittest.TestCase):
         """from_amplifyメソッドのテスト"""
         # amplifyの設定
         q = VariableGenerator().array("Binary", shape=(3))
-        objective = q[0] * q[1] - q[2]
+        objective = q[0] * q[1] - q[2] + 20
 
         # QuData.inputオブジェクトを作成し、amplify問題を渡す
         qudata = QuData.input().from_amplify(objective)
@@ -477,6 +481,18 @@ class TestQudata(unittest.TestCase):
         self.assertEqual(qdo.result_type, 'pulp')
 
     def test_from_amplify(self):
+        """Amplify形式の結果を受け取るメソッドのテスト"""
+        try:
+            from amplify import VariableGenerator, Model, FixstarsClient, solve
+        except Exception:
+            self.skipTest(
+                "Amplifyがインストールされていないためテストをスキップします。"
+            )
+
+        from amplify import VariableGenerator, Model, FixstarsClient, solve
+        from datetime import timedelta
+        import os
+
         # Amplify形式の結果
         gen = VariableGenerator()
         q = gen.array("Binary", shape=(3))
@@ -484,7 +500,12 @@ class TestQudata(unittest.TestCase):
 
         # ソルバーの設定
         client = FixstarsClient()
-        client.token = "AE/HaqGh1iuFMEennXk10xS1LCgld8D18oC"
+        client.token = os.getenv("AMPLIFY_TOKEN")
+        if client.token is None:
+            self.skipTest(
+                "Amplifyのトークンが設定されていないためテストをスキップします。"
+            )
+
         client.parameters.timeout = timedelta(milliseconds=100)
 
         # 最小化を実行
@@ -552,12 +573,12 @@ class TestQudata(unittest.TestCase):
 
         # 検証
         self.assertEqual(qdo.result, expected_result)
-        self.assertEqual(qdo.result_type, 'sympy')
+        self.assertEqual(qdo.result_type, 'scipy')
 
     def test_to_dimod(self):
         # Dimod形式の結果に変換
         qdo = QuData.output(
-            result={'variables': {'q0': 0.0, 'q1': 1.0, 'q2': 1.0}, 'objective': -2}
+            result={'solution': {'q0': 0.0, 'q1': 1.0, 'q2': 1.0}, 'energy': -2}
         )
         dimod_result = qdo.to_dimod()
 
@@ -571,7 +592,7 @@ class TestQudata(unittest.TestCase):
     def test_to_scipy(self):
         # SciPy形式の結果に変換
         qdo = QuData.output(
-            result={'variables': {'q0': 0.0, 'q1': 1.0, 'q2': 1.0}, 'objective': -2}
+            result={'solution': {'q0': 0.0, 'q1': 1.0, 'q2': 1.0}, 'energy': -2}
         )
         scipy_result = qdo.to_scipy()
 
@@ -600,4 +621,5 @@ class TestQudata(unittest.TestCase):
 
 
 if __name__ == '__main__':
-    unittest.main()
+    runner = unittest.TextTestRunner(verbosity=2)
+    unittest.main(testRunner=runner)